[FE] ♻️ Refactor: 방명록 코드 변경 수정 및 파일이동 파일 삭제  #90
